재고 차감 로직 인터페이스 분리 #29 TicketStockService 인터페이스 정의 및 비관적 락 구현 클래스 추가.
