+++ src/main/resources/application.yml
@@ -7,3 +7,7 @@ management:
     web:
       exposure:
         include: health,metrics,prometheus
+
+ticket:
+  lock:
+    strategy: pessimistic
